refactoring shorten array function
I went through and reevaluated the shorten functions that existed in the program. I had to redo one of the older versions of it in order to handle large changes with parentheticles. The new method allows variables to be bypassed when present in the functions (still requires some work though).